+++ CHANGELOG.md
@@ -15,6 +15,10 @@ You can find its changes [documented below](#070---2026-01-13).
 
 This release has an [MSRV][] of 1.92.
 
+### Added
+
+- Image tinting with alpha-mask and component-wise multiply modes.
+
 ## [0.7.0][] - 2026-01-13
 
 This release has an [MSRV][] of 1.92.